requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.3
decorator==4.0.11
enum34==1.1.6
google-api-python-client==1.6.2
gspread==0.6.2
httplib2==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
oauth2client==4.0.0
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
Pygments==2.2.0
pyparsing==2.2.0
-e git+https://github.com/Nayigiziki/python-craigslist.git@ad30b3640fad5e8d1442f6adc2bb6b4fd372cd6a#egg=python_craigslist
python-dateutil==2.6.0
requests==2.13.0
rsa==3.4.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.5
SQLAlchemy==1.1.9
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
websocket-client==0.40.0
